Move `getLinkedEntity` and do other minor utility clean-ups
