+++ src/h5web/visualizations/nexus/utils.ts
@@ -5,10 +5,10 @@ import {
   HDF5Value,
   HDF5Entity,
 } from '../../providers/models';
-import { assertGroup, getEntity, isGroup } from '../../providers/utils';
+import { assertGroup, getLinkedEntity, isGroup } from '../../providers/utils';
 import {
   NxAttribute,
-  NX_INTERPRETATIONS,
+  NxInterpretation,
   RawSilxStyle,
   SilxStyle,
 } from './models';
@@ -19,44 +19,11 @@ import {
   isScaleType,
 } from '../shared/utils';
 
-export function isNxInterpretation(attrValue: HDF5Value): boolean {
-  return (
-    attrValue &&
-    typeof attrValue === 'string' &&
-    NX_INTERPRETATIONS.includes(attrValue)
-  );
-}
-
 export function getAttributeValue(
   entity: HDF5Dataset | HDF5Group,
   attributeName: NxAttribute
 ): HDF5Value | undefined {
-  if (!entity.attributes) {
-    return undefined;
-  }
-
-  return entity.attributes.find((attr) => attr.name === attributeName)?.value;
-}
-
-export function getLinkedEntity(
-  entityName: string,
-  group: HDF5Group,
-  metadata: HDF5Metadata
-): HDF5Entity | undefined {
-  const childLink = group.links?.find((l) => l.title === entityName);
-  return getEntity(childLink, metadata);
-}
-
-export function getNxAxisNames(group: HDF5Group): (string | undefined)[] {
-  const axisList = getAttributeValue(group, 'axes');
-
-  if (!axisList) {
-    return [];
-  }
-
-  const axisNames = typeof axisList === 'string' ? [axisList] : axisList;
-  assertArray<string>(axisNames);
-  return axisNames.map((a) => (a !== '.' ? a : undefined));
+  return entity.attributes?.find((attr) => attr.name === attributeName)?.value;
 }
 
 export function findNxDataGroup(
@@ -90,6 +57,26 @@ export function findNxDataGroup(
   assertGroup(defaultEntity, `Expected group at path "${defaultPath}"`);
 
   return findNxDataGroup(defaultEntity, metadata);
+}
+
+export function isNxInterpretation(
+  attrValue: HDF5Value
+): attrValue is NxInterpretation {
+  return (
+    typeof attrValue === 'string' &&
+    Object.values<string>(NxInterpretation).includes(attrValue)
+  );
+}
+
+export function getNxAxisNames(group: HDF5Group): (string | undefined)[] {
+  const axisList = getAttributeValue(group, 'axes');
+  if (!axisList) {
+    return [];
+  }
+
+  const axisNames = typeof axisList === 'string' ? [axisList] : axisList;
+  assertArray<string>(axisNames);
+  return axisNames.map((a) => (a !== '.' ? a : undefined));
 }
 
 export function getDatasetLabel(
